Refactoring: GridCreatingConditionTriangle::AddPolygonVertexCommand implementation is separated from gridcreatingconditiontriangle.cpp.
